merge-coverage-report script and hook
